use smaller type, may fit more into on-stack vector

+++ arangod/Aql/Expression.cpp
@@ -913,8 +913,8 @@ AqlValue Expression::executeSimpleExpressionFCall(
   VPackFunctionParameters parameters{arena};
  
   // same here
-  SmallVector<uint64_t>::allocator_type::arena_type arena2;
-  SmallVector<uint64_t> destroyParameters{arena2};
+  SmallVector<uint8_t>::allocator_type::arena_type arena2;
+  SmallVector<uint8_t> destroyParameters{arena2};
   parameters.reserve(n);
   destroyParameters.reserve(n);
